requirements.txt
asgiref
attrs
autobahn
Automat
beautifulsoup4
bs4
cachetools
certifi
cffi
cfgv
charset-normalizer
constantly
coreapi
coreschema
cryptography
daphne
decouple
defusedxml
distlib
Django
django-rest-auth
django-rest-swagger
djangorestframework
djangorestframework-simplejwt
drf-standardized-errors
drf-yasg
filelock
google
google-auth
gunicorn
hiredis
httplib2
hyperlink
identify
idna
incremental
inflection
isort
itypes
Jinja2
MarkupSafe
msgpack
multidict
nodeenv
oauthlib
openapi-codec
packaging
pillow
platformdirs
pre-commit
pyasn1
pyasn1_modules
pycparser
pydub
PyJWT
pyOpenSSL
pyparsing
python-dotenv
python3-openid
pytz
PyYAML
requests
requests-oauthlib
rest-framework-simplejwt
rsa
service-identity
setuptools
simplejson
six
soupsieve
sqlparse
Twisted
txaio
typing_extensions
uritemplate
urllib3
virtualenv
whitenoise
zope.interface
